1.54 fixed bugs of DatumFloat, Schema.indexMap(all key use toUpperCase())
Former-commit-id: 9ef36680f21c3d070fc8c5b5ca98dd0d918d8957

+++ src/dao/DatumFloat.java
@@ -25,7 +25,7 @@ public class DatumFloat extends Datum{
 
 	@Override
 	public void setNumericValue(double valueIn) {
-		value = (int)valueIn;
+		value = (float)valueIn;
 	}
 	
 	@Override
